# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "B2B Marketplace API",
    "description": "API for B2B Marketplace Platform",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server (HTTP)"
    },
    {
      "url": "https://localhost:5001",
      "description": "Development server (HTTPS)"
    }
  ],
  "paths": {
...
    "/api/admin/certifications/{id}": {
      "get": {
        "tags": [
          "AdminCertification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AdminCertification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AdminCertification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Certifications": {
      "post": {
        "tags": [
          "Certifications"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "ContentType": {
                    "type": "string"
                  },
                  "ContentDisposition": {
                    "type": "string"
                  },
                  "Headers": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "Length": {
                    "type": "integer",
                    "format": "int64"
                  },
                  "Name": {
                    "type": "string"
                  },
                  "FileName": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "name": {
                  "style": "form"
                },
                "ContentType": {
                  "style": "form"
                },
                "ContentDisposition": {
                  "style": "form"
                },
                "Headers": {
                  "style": "form"
                },
                "Length": {
                  "style": "form"
                },
                "Name": {
                  "style": "form"
                },
                "FileName": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Certifications/{id}/status": {
      "put": {
        "tags": [
          "Certifications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.UpdateCertificationStatusDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.UpdateCertificationStatusDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.UpdateCertificationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Certifications/{id}": {
      "get": {
        "tags": [
          "Certifications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "B2BMarketplace.Core.DTOs.CertificationDto": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Status": {
            "$ref": "#/components/schemas/B2BMarketplace.Core.Enums.CertificationStatus"
          },
          "SubmittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "ReviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "AdminNotes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "B2BMarketplace.Core.DTOs.UpdateCertificationStatusDto": {
        "required": [
          "Status"
        ],
        "type": "object",
        "properties": {
          "Status": {
            "$ref": "#/components/schemas/B2BMarketplace.Core.Enums.CertificationStatus"
          },
          "AdminNotes": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "B2BMarketplace.Core.Enums.CertificationStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}